// ux/api/client/types.ts
export interface components {
    schemas: {
        /** DomainCreate */
        DomainCreate: {
            /** Title */
            title: string;
            /** Definition */
            definition: string;
            /**
             * Layer Id
             * Format: uuid
             */
            layer_id: string;
        };
        /** DomainOut */
        DomainOut: {
            /** Title */
            title: string;
            /** Definition */
            definition: string;
            /**
             * Layer Id
             * Format: uuid
             */
            layer_id: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Title Embedding */
            title_embedding?: number[] | null;
            /** Definition Embedding */
            definition_embedding?: number[] | null;
            /** Created At */
            created_at: string;
            /** Version */
            version?: number | null;
            /** Last Modified */
            last_modified?: string | null;
        };
// ...
        LayerOut: {
            /** Title */
            title: string;
            /** Definition */
            definition?: string | null;
            /** Primary Predicate */
            primary_predicate?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Title Embedding */
            title_embedding?: number[] | null;
            /** Definition Embedding */
            definition_embedding?: number[] | null;
            /** Created At */
            created_at: string;
            /** Version */
            version?: number | null;
            /** Last Modified */
            last_modified?: string | null;
        };
// ...
        TermOut: {
            /** Title */
            title: string;
            /** Definition */
            definition: string;
            /**
             * Domain Id
             * Format: uuid
             */
            domain_id: string;
            /**
             * Layer Id
             * Format: uuid
             */
            layer_id: string;
            /** Parent Term Id */
            parent_term_id?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Title Embedding */
            title_embedding?: number[] | null;
            /** Definition Embedding */
            definition_embedding?: number[] | null;
            /** Created At */
            created_at: string;
            /** Version */
            version: number;
            /** Last Modified */
            last_modified: string;
        };
// ...
        TermRelationshipOut: {
            /**
             * Source Term Id
             * Format: uuid
             */
            source_term_id: string;
            /**
             * Target Term Id
             * Format: uuid
             */
            target_term_id: string;
            /** Predicate */
            predicate: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Created At */
            created_at: string;
            /** Source Term */
            source_term?: {
                [key: string]: unknown;
            } | null;
            /** Target Term */
            target_term?: {
                [key: string]: unknown;
            } | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}